requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
torch==2.2.2
torchaudio==2.2.2
pyannote.audio==3.1.1
numpy==1.26.4
